Add config to turn off auto tag sync with Crate by setting SKIP_CRATE_TAG_AUTOSYNC to true

+++ opentaps_seas/core/models.py
@@ -416,14 +416,16 @@ def entity_deleted(sender, instance, using, **kwargs):
     logger.info('entity_deleted: %s', instance.entity_id)
     EntityNote.objects.filter(entity_id=instance.entity_id).delete()
     EntityFile.objects.filter(entity_id=instance.entity_id).delete()
-    delete_tags_from_crate_entity(instance)
+    if not settings.SKIP_CRATE_TAG_AUTOSYNC:
+        delete_tags_from_crate_entity(instance)
 
 
 @receiver(post_save, sender=Entity, dispatch_uid='entity_post_save_signal')
 def entity_saved(sender, instance, using, **kwargs):
     # remove all associated resourcese: notes, files, links ...
     logger.info('entity_saved: %s', instance.entity_id)
-    sync_tags_to_crate_entity(instance)
+    if not settings.SKIP_CRATE_TAG_AUTOSYNC:
+        sync_tags_to_crate_entity(instance)
 
 
 class Topic(models.Model):
